requirements.txt
numpy>=1.26
pandas>=2.1
scikit-learn>=1.4
scipy>=1.11
joblib>=1.3
requests>=2.31
beautifulsoup4>=4.12
python-dateutil>=2.8
pyyaml>=6.0
